backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-socketio
pyjwt
pymongo